+++ src/current-refinements/current-refinements.ts
@@ -1,7 +1,7 @@
 import { Component, Input, Inject, forwardRef } from "@angular/core";
 
 import { connectCurrentRefinedValues } from "instantsearch.js/es/connectors";
-import { noop, isFunction } from "lodash-es";
+import { capitalize, noop, isFunction } from "lodash-es";
 
 import { BaseWidget } from "../base-widget";
 import { NgAisInstantSearch } from "../instantsearch/instantsearch";
@@ -29,22 +29,20 @@ export type CurrentRefinementsState = {
         {{clearRefinementsLabel}}
       </button>
 
-      <ul [class]="cx('list')">
-        <li
-          [class]="cx('item')"
-          *ngFor="let refinement of refinements"
-          (click)="handleClick($event, refinement)"
-        >
-          <button [class]="cx('button')">
-            <span [class]="cx('label')">{{refinement.computedLabel}}</span>
-            <span
-              *ngIf="refinement.count && refinement.count > 0"
-              [class]="cx('count')"
-            >
-              {{refinement.count}}
-            </span>
-            <span [class]="cx('delete')">✕</span>
-          </button>
+      <ul
+        [class]="cx('list')"
+        *ngFor="let refinement of refinements"
+      >
+        <li [class]="cx('item')">
+          <span [class]="cx('label')">{{refinement.label}}:</span>
+
+          <span
+            [class]="cx('category')"
+            *ngFor="let item of refinement.items"
+          >
+            <span [class]="cx('categoryLabel')">{{item.name}}</span>
+            <button [class]="cx('delete')" (click)="handleClick($event, item)">✕</button>
+          </span>
         </li>
       </ul>
 
@@ -86,9 +84,31 @@ export class NgAisCurrentRefinements extends BaseWidget {
   }
 
   get refinements() {
-    return isFunction(this.transformItems)
+    const items = isFunction(this.transformItems)
       ? this.transformItems(this.state.refinements)
       : this.state.refinements;
+
+    // group refinements by category? (attributeName && type)
+    return items.reduce((res, { type, attributeName, ...refinement }) => {
+      const match = res.find(
+        r => r.attributeName === attributeName && r.type === type
+      );
+      if (match) {
+        match.items.push({ type, attributeName, ...refinement });
+      } else {
+        res.push({
+          type,
+          attributeName,
+          label: capitalize(attributeName),
+          items: [{ type, attributeName, ...refinement }]
+        });
+      }
+      return res;
+    }, []);
+  }
+
+  get json() {
+    return JSON.stringify(this.refinements, null, 4);
   }
 
   constructor(